scrapers/STRATAXIS-RENT PRICE-SCRAPER/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
python-dateutil==2.8.2
tqdm==4.66.1
fake-useragent==1.4.0
selenium==4.16.0
validators==0.22.0
